wordpress-telegram-bot/requirements.txt
python-telegram-bot>=20.3
aiohttp>=3.8.4
python-dotenv>=1.0.0
Pillow>=9.5.0
requests>=2.31.0
python-magic>=0.4.27
